requirements.txt
streamlit 
pydantic 
python-decouple